+++ src/Echelon.Web/Controllers/TasksController.cs
@@ -1,9 +1,9 @@
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
 using System.ComponentModel.DataAnnotations;
 using Echelon.Application.DTOs;
 using Echelon.Application.Services;
 using Echelon.Infrastructure.Auth;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Echelon.Web.Controllers;